+++ frontend/src/pages/projectDetail/projectDetail.gql
